engine: full interfaces implementation
