--- internal/comment/helper.go
@@ -1,42 +1,0 @@
-package comment
-
-import (
-	"context"
-	"database/sql"
-	"fmt"
-	"strings"
-
-	"godiscourse/internal/durable"
-
-	"github.com/gofrs/uuid"
-)
-
-var commentColumns = []string{"comment_id", "body", "topic_id", "user_id", "score", "created_at", "updated_at"}
-
-func (c *Model) values() []interface{} {
-	return []interface{}{c.CommentID, c.Body, c.TopicID, c.UserID, c.Score, c.CreatedAt, c.UpdatedAt}
-}
-
-func commentFromRows(row durable.Row) (*Model, error) {
-	var c Model
-	err := row.Scan(&c.CommentID, &c.Body, &c.TopicID, &c.UserID, &c.Score, &c.CreatedAt, &c.UpdatedAt)
-	return &c, err
-}
-
-func findComment(ctx context.Context, tx *sql.Tx, id string) (*Model, error) {
-	if _, err := uuid.FromString(id); err != nil {
-		return nil, nil
-	}
-	row := tx.QueryRowContext(ctx, fmt.Sprintf("SELECT %s FROM comments WHERE comment_id=$1", strings.Join(commentColumns, ",")), id)
-	c, err := commentFromRows(row)
-	if err == sql.ErrNoRows {
-		return nil, nil
-	}
-	return c, err
-}
-
-func commentsCountByTopic(ctx context.Context, tx *sql.Tx, id string) (int64, error) {
-	var count int64
-	err := tx.QueryRowContext(ctx, "SELECT count(*) FROM comments WHERE topic_id=$1", id).Scan(&count)
-	return count, err
-}
