+++ internal/controllers/category.go
@@ -39,7 +39,7 @@ func (impl *categoryImpl) topics(w http.ResponseWriter, r *http.Request, params 
 		views.RenderErrorResponse(w, r, err)
 	} else if category == nil {
 		views.RenderErrorResponse(w, r, session.NotFoundError(r.Context()))
-	} else if topics, err := impl.poster.GetTopicByCategoryID(r.Context(), category.CategoryID, offset); err != nil {
+	} else if topics, err := impl.poster.GetTopicsByCategoryID(r.Context(), category.CategoryID, offset); err != nil {
 		views.RenderErrorResponse(w, r, err)
 	} else {
 		views.RenderTopics(w, r, topics)
